+++ templates/crm/emails/user_credentials.html
@@ -34,7 +34,7 @@
 
             <p>Aqui estão suas credenciais de acesso ao portal de edição de templates:</p>
             <div class="credentials">
-                <p><strong>Usuário (Email):</strong> {{ usuario_email }}</p>
+                <p><strong>Username:</strong> {{ usuario_username }}</p>
                 <p><strong>Senha:</strong> {{ usuario_senha }}</p>
             </div>
